Simplify TransactionPlanExecutor by removing factories
